Add files via upload
Numerous changes across the php and js files -- everything worked on localhost -- when I uploaded to GoDaddy all of a sudden the mysql broke and I could not initialize the DB.  The stored function syntax needed to be rewritten.  Other things broke.  Many changes -- a whole day of work.  I think I got it now.