Add files via upload
Tag Cloud Search Bricks are now clickable and searchable from anywhere in the UI.  That includes the returned HTML blog posts that are tagged with bricks.  These bricks can now be added/removed from a search and will be updated when search bricks change.  updated everything including the README and graphics.